The webhook conformance tests get their own file
Main is red on the craft gate: the webhooks integration file reached 1004
lines, four past the cap, and craft blocks a MAJOR.

The two schema-conformance tests move out. They are a different question
from the rest of that file — those ask whether a webhook arrives, retries
and stops; these ask whether what arrived is the shape the contract
promised, re-deriving the schema from public-events.yaml rather than from
the generated Go types.

Both pass in the integration lane after the move, and the remaining file
is 870 lines.
